[ REFACTOR ] Alerando nome do componente CarregandoVuetify
