+++ front/src/modules/projeto/visualizar/components/execucao/components/VisualizarFiscalizacao.vue
@@ -601,7 +601,7 @@
 </template>
 <script>
     import { mapActions, mapGetters } from 'vuex';
-    import Carregando from '@/components/Carregando_vuetify';
+    import Carregando from '@/components/CarregandoVuetify';
 
     export default {
         name: 'VisualizarFiscalizacao',
