+++ front/src/modules/projeto/visualizar/components/outrasInformacoes/CertidoesNegativas.vue
@@ -31,7 +31,7 @@
 <script>
 
     import { mapActions, mapGetters } from 'vuex';
-    import Carregando from '@/components/Carregando_vuetify';
+    import Carregando from '@/components/CarregandoVuetify';
 
     export default {
         name: 'CertidoesNegativas',
