+++ front/src/modules/projeto/visualizar/components/outrasInformacoes/DocumentosAnexados.vue
@@ -46,7 +46,7 @@
 </template>
 <script>
     import { mapGetters, mapActions } from 'vuex';
-    import Carregando from '@/components/Carregando_vuetify';
+    import Carregando from '@/components/CarregandoVuetify';
 
     export default {
         name: 'DocumentosAnexados',
