+++ front/src/modules/projeto/visualizar/components/outrasInformacoes/DocumentosAssinados.vue
@@ -52,7 +52,7 @@
 
 <script>
     import { mapActions, mapGetters } from 'vuex';
-    import Carregando from '@/components/Carregando_vuetify';
+    import Carregando from '@/components/CarregandoVuetify';
 
     export default {
         name: 'DocumentosAssinados',
