+++ front/src/modules/projeto/visualizar/components/outrasInformacoes/HistoricoEncaminhamento.vue
@@ -30,7 +30,7 @@
 </template>
 <script>
     import { mapActions, mapGetters } from 'vuex';
-    import Carregando from '@/components/Carregando_vuetify';
+    import Carregando from '@/components/CarregandoVuetify';
 
     export default {
         name: 'HistoricoEncaminhamento',
