+++ front/src/modules/projeto/visualizar/components/outrasInformacoes/LocalRealizacaoDeslocamento.vue
@@ -60,7 +60,7 @@
 </template>
 <script>
     import { mapActions, mapGetters } from 'vuex';
-    import Carregando from '@/components/Carregando_vuetify';
+    import Carregando from '@/components/CarregandoVuetify';
 
     export default {
         name: 'LocalRealizacaoDeslocamento',
