+++ front/src/modules/projeto/visualizar/components/outrasInformacoes/PlanoDistribuicaoIn2013.vue
@@ -133,7 +133,7 @@
 </template>
 <script>
     import { mapGetters, mapActions } from 'vuex';
-    import Carregando from '@/components/Carregando_vuetify';
+    import Carregando from '@/components/CarregandoVuetify';
 
     export default {
         name: 'PlanoDistribuicao-in-2013',
