+++ front/src/modules/projeto/visualizar/components/outrasInformacoes/ProvidenciaTomada.vue
@@ -36,7 +36,7 @@
 
 <script>
     import { mapActions, mapGetters } from 'vuex';
-    import Carregando from '@/components/Carregando_vuetify';
+    import Carregando from '@/components/CarregandoVuetify';
     import cnpjFilter from '@/filters/cnpj';
 
     export default {
