+++ front/src/modules/projeto/visualizar/components/outrasInformacoes/components/VisualizarDiligenciaAdequacao.vue
@@ -65,7 +65,7 @@
 
 <script>
     import { mapActions, mapGetters } from 'vuex';
-    import Carregando from '@/components/Carregando_vuetify';
+    import Carregando from '@/components/CarregandoVuetify';
 
     export default {
         name: 'VisualizarDiligenciaAdequacao',
