+++ front/src/modules/projeto/visualizar/components/outrasInformacoes/components/VisualizarDiligenciaProposta.vue
@@ -85,7 +85,7 @@
 
 <script>
     import { mapActions, mapGetters } from 'vuex';
-    import Carregando from '@/components/Carregando_vuetify';
+    import Carregando from '@/components/CarregandoVuetify';
 
     export default {
         name: 'VisualizarDiligenciaProposta',
